[REFACTOR] import 경로 변경 문제 해결 #82
